requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0

# 文档解析
pypdf2==3.0.1
pdfminer.six==20221105
beautifulsoup4==4.12.2
html2text==2020.1.16

# AI集成
langchain==0.0.335
openai==1.3.5
tiktoken==0.5.1

# 数据库
motor==3.3.1
pymongo==4.6.0
redis==5.0.1

# 任务队列
celery==5.3.4
flower==2.0.1

# 通知系统
fastapi-mail==1.4.1
aioredis==2.0.1
websockets==12.0

# 测试
pytest==7.4.3
httpx==0.25.1 